Stuff done
- fixed graphical bugs w/ weapon FX/animations
- finally added actual loot and text to show loot
- added text to show EXP and silver gain
- implemented leveling up
- town to dungeon --- players set to inactive
- dungeon to battle --- players set to active
- battle to dungeon --- players set to inactive

+++ Assets/Scripts/ScriptableObjects/EnemyEncounter.cs
@@ -31,8 +31,8 @@ public class EnemyEncounter : ScriptableObject
         return silver;
     }
 
-    public int GetExperience(){
-        int exp = 0;
+    public float GetExperience(){
+        float exp = 0;
         foreach(GameObject enemy in enemies){
             CharacterSheet charSheet = enemy.GetComponent<CharacterSheet>();
             exp += charSheet.GetExperience();
